Exceções e ControllerExceptionHandler

+++ src/main/java/com/baggio/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.baggio.dscommerce.services;
 import com.baggio.dscommerce.dto.ProductDTO;
 import com.baggio.dscommerce.entities.Product;
 import com.baggio.dscommerce.repositories.ProductRepository;
+import com.baggio.dscommerce.services.exceptions.DatabaseException;
+import com.baggio.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,7 +25,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product =  productRepository.findById(id).get();
+        Product product =  productRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado com id: " + id));
         return new  ProductDTO(product);
     }
 
@@ -41,15 +47,26 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product product = productRepository.getReferenceById(id);
-        dtoToProduct(productDTO, product);
-        product = productRepository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            dtoToProduct(productDTO, product);
+            product = productRepository.save(product);
+            return new ProductDTO(product);
+        }catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado com id: " + id);
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado com id: " + id);
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void dtoToProduct(ProductDTO productDTO, Product product) {
